#4: Prepare the project for release * project separated into modules: core, slick, examples * added workflows * removed wrong packages that remained after previous merge
