+++ .github/WORKFLOWS/build.yml
@@ -13,23 +13,30 @@
 # limitations under the License.
 #
 
-name: License Check
+name: Build
 
 on:
   push:
-    branches: [ main, develop, master ]
+    branches: [master ]
   pull_request:
-    branches: [ master ]
-    types: [ assigned, opened, synchronize, reopened, labeled ]
+      branches: [ master ]
+      types: [ assigned, opened, synchronize, reopened, labeled ]
 
 jobs:
-  license-test:
+  test:
     runs-on: ubuntu-latest
+    strategy:
+      fail-fast: false
+      matrix:
+        scala: [2.11.12, 2.12.12]
+    name: Build on Scala ${{matrix.scala}}
     steps:
       - name: Checkout code
         uses: actions/checkout@v2
+      - uses: coursier/cache-action@v5
       - name: Setup Scala
         uses: olafurpg/setup-scala@v10
         with:
           java-version: "adopt@1.8"
-      - run: sbt headerCheck
+      - name: Build and run tests
+        run: sbt ++${{matrix.scala}} test
