+++ core/src/main/scala/za/co/absa/fadb/exceptions/DBFailException.scala
