+++ core/src/main/scala/za/co/absa/fadb/naming_conventions/NamingConvention.scala
