+++ core/src/main/scala/za/co/absa/fadb/naming_conventions/letters_case/LettersCase.scala
