+++ core/src/test/scala/za/co/absa/fadb/naming_conventions/ExplicitNamingRequiredSuite.scala
