+++ project/plugins.sbt
@@ -1,4 +1,3 @@
-
 /*
  * Copyright 2019 ABSA Group Limited
  *
@@ -14,6 +13,6 @@
  * limitations under the License.
  */
 
-addSbtPlugin("de.heikoseeberger" % "sbt-header" % "5.6.0")
-addSbtPlugin("com.github.sbt" % "sbt-release" % "1.1.0")
-addSbtPlugin("com.github.sbt" % "sbt-pgp" % "2.1.2")
+addSbtPlugin("de.heikoseeberger"  % "sbt-header"   % "5.6.0")
+addSbtPlugin("com.github.sbt"     % "sbt-release"  % "1.1.0")
+addSbtPlugin("com.github.sbt"     % "sbt-pgp"      % "2.1.2")
