--- src/main/scala/za/co/absa/faDB/namingConventions/AsIsNaming.scala
@@ -1,31 +1,0 @@
-/*
- * Copyright 2022 ABSA Group Limited
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package za.co.absa.faDB.namingConventions
-
-import za.co.absa.faDB.namingConventions.lettersCase.LettersCase
-import za.co.absa.faDB.namingConventions.lettersCase.LettersCase.AsIs
-
-class AsIsNaming(lettersCase: LettersCase) extends NamingConvention{
-  override def stringPerConvention(original: String): String = {
-    lettersCase.convert(original)
-  }
-}
-
-object AsIsNaming {
-  object Implicits {
-    implicit val namingConvention: NamingConvention = new AsIsNaming(AsIs)
-  }
-}
