--- src/test/scala/za/co/absa/faDB/enceladus/DatasetSchemaSuite.scala
@@ -1,129 +1,0 @@
-/*
- * Copyright 2021 ABSA Group Limited
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package za.co.absa.faDB.enceladus
-
-import org.scalatest.funsuite.AnyFunSuite
-import org.scalatest.matchers.should.Matchers
-import org.scalatest.wordspec.AnyWordSpec
-import za.co.absa.faDB.DBSession
-import za.co.absa.faDB.enceladus.DatasetSchema.{Schema, SchemaInput}
-import za.co.absa.faDB.exceptions.DBFailException
-
-import scala.concurrent.Await
-import scala.concurrent.duration.Duration
-
-class DatasetSchemaSuite extends AnyWordSpec with Matchers {
-  private val session = new DBSession("menasdb")
-  private val schemas = new DatasetSchema(session)
-
-  private def checkException(exception: DBFailException): Unit = {
-    println(s"Requested failed with: ${exception.status} - ${exception.message}")
-  }
-
-  "listSchemas" should {
-    "list the schemas" in {
-      val ls = schemas.listSchemas()
-      val result = Await.result(ls, Duration.Inf)
-      result.foreach(println)
-    }
-  }
-
-  "getSchema" should {
-    "return the particular schema" when {
-      "given name and version" in {
-        val ls = schemas.getSchema(("aaa", Option(1)))
-        val result = Await.result(ls, Duration.Inf)
-        println(result)
-      }
-      "given id" in {
-        val gs = schemas.getSchema(1000000000000051L)
-        val result = Await.result(gs, Duration.Inf)
-        println(result)
-      }
-    }
-    "return the latest schema version" when {
-      "only the schema name is given" in {
-        val ls = schemas.getSchema(("aaa", None))
-        val result = Await.result(ls, Duration.Inf)
-        println(result)
-      }
-    }
-    "fail" when {
-      "schema does not exist" in {
-        val exception = intercept[DBFailException] {
-          val gs = schemas.getSchema(("xxx", None))
-          Await.result(gs, Duration.Inf)
-        }
-        checkException(exception)
-      }
-      "requested schema version does not exist" in {
-        val exception = intercept[DBFailException] {
-          val gs = schemas.getSchema(("aaa", Some(1000)))
-          Await.result(gs, Duration.Inf)
-        }
-        checkException(exception)
-      }
-    }
-  }
-
-  "addSchema" should {
-    "add a schema" in {
-      val schemaInput = SchemaInput(
-        schemaName = "bbe",
-        schemaVersion = 1,
-        schemaDescription = Option("Hello World"),
-        fields = Option("""{"lorem": "ipsum"}"""),
-        userName = "david"
-      )
-      val result = Await.result(schemas.addSchema(schemaInput), Duration.Inf)
-      println(result)
-    }
-    "fail" when {
-      "Schema already exists" in {
-        val schemaInput = SchemaInput(
-          schemaName = "aaa",
-          schemaVersion = 2,
-          schemaDescription = Option("Updates"),
-          fields = Option("""{"foo": "bar"}"""),
-          userName = "david"
-        )
-        val exception = intercept[DBFailException] {
-          Await.result(schemas.addSchema(schemaInput), Duration.Inf)
-        }
-        checkException(exception)
-      }
-      "Schema version wrong" in {
-        val schemaInput = SchemaInput(
-          schemaName = "aaa",
-          schemaVersion = 1000,
-          schemaDescription = Option("Will fail"),
-          fields = Option("""{"not_getting_in": "1"}"""),
-          userName = "david"
-        )
-        val exception = intercept[DBFailException] {
-          Await.result(schemas.addSchema(schemaInput), Duration.Inf)
-        }
-        checkException(exception)
-      }
-      "Schema has been disabled" in {
-        // TODO
-      }
-      "Schema has been locked" in {
-        // TODO
-      }
-    }
-  }
-}
