--- src/test/scala/za/co/absa/faDB/examples/ExampleSuite.scala
@@ -1,28 +1,0 @@
-/*
- * Copyright 2021 ABSA Group Limited
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package za.co.absa.faDB.examples
-
-import org.scalatest.funsuite.AnyFunSuite
-
-class ExampleSuite extends AnyFunSuite {
-  test("foo") {
-//    val expected = Seq(72, 101, 108, 108, 111)
-//    val session = new DBSession
-//    val example = new Example(session)
-//    val result = example.myFunction("Hello")
-//    assert(result == expected)
-  }
-}
